requirements.txt
Django>=3.1,<3.2
djangorestframework>=3.12,<3.13
markdown>=3.3,<3.4
django-filter>=2.4,<2.5
djoser>=2.1,<2.2
djangorestframework_simplejwt>=4.6,<4.7
django-environ>=0.4,<0.5
redis>=3.5,<3.6
django-redis>=4.12,<4.13
django-redisboard>=4.1,<4.2
django-extensions>=3.1.1,<3.2
drf-yasg>=1.20,<1.21
celery>=5.0,<5.1
django-celery-beat>=2.2,<2.3
Pygments>=2.8.0,<2.9
psycopg2-binary>=2.8.6,<2.9
ipython>=7.21.0,<7.22
django-taggit>=1.3,<1.4
